+++ apps/mobile/src/navigations/bottom-tabs-navigator/bottom-tabs-navigator.tsx
@@ -9,12 +9,14 @@ import {
 	GradientTabIcon,
 	Text,
 } from "~/libs/components/components";
-import { BaseColor, BottomTabScreenName } from "~/libs/enums/enums";
+import { BaseColor, BottomTabScreenName, DataStatus } from "~/libs/enums/enums";
+import { useAppDispatch, useAppSelector, useEffect } from "~/libs/hooks/hooks";
 import { type BottomTabNavigationParameterList } from "~/libs/types/types";
 import { WheelStackNavigator } from "~/navigations/bottom-tabs-navigator/wheel/wheel";
 import { Chat } from "~/screens/chat/chat";
 import { Settings } from "~/screens/settings/settings";
 import { Tasks } from "~/screens/tasks/tasks";
+import { actions as appActions } from "~/slices/app/app";
 
 import { styles } from "./styles";
 
@@ -36,9 +38,31 @@ const screenOptions: BottomTabNavigationOptions = {
 };
 
 const BottomTabsNavigator: React.FC = () => {
+	const dispatch = useAppDispatch();
+	const dataStatus = useAppSelector(({ app }) => app.dataStatus);
+	const initialNotificationId = useAppSelector(
+		({ app }) => app.initialNotificationId,
+	);
+	const isOpenedWithExpiredTaskNotification =
+		initialNotificationId === "default";
+
+	const isLoading = dataStatus === DataStatus.PENDING;
+
+	useEffect(() => {
+		void dispatch(appActions.updateInitialNotificationId());
+	}, [dispatch]);
+
+	if (isLoading) {
+		return null;
+	}
+
 	return (
 		<BottomTabs.Navigator
-			initialRouteName={BottomTabScreenName.WHEEL}
+			initialRouteName={
+				isOpenedWithExpiredTaskNotification
+					? BottomTabScreenName.TASKS
+					: BottomTabScreenName.WHEEL
+			}
 			screenOptions={screenOptions}
 		>
 			<BottomTabs.Screen