+++ apps/mobile/src/screens/chat/libs/components/chat-buttons/libs/components/components.ts
@@ -1,3 +1,4 @@
-export { AcceptTasks } from "./accept-tasks/accept-tasks";
 export { CheckBoxButtons } from "./checkbox-buttons/checkbox-buttons";
 export { GenerateTaskButtons } from "./generate-task-buttons/generate-task-buttons";
+export { AcceptRegenerateTaskButtons } from "~/screens/chat/libs/components/chat-buttons/libs/components/accept-regenerate-task-buttons/accept-regenerate-task-buttons";
+export { ExplainAcceptTaskButtons } from "~/screens/chat/libs/components/chat-buttons/libs/components/explain-accept-task-buttons/explain-accept-task-buttons";