+++ apps/mobile/src/screens/notification-questions/libs/types/types.ts
@@ -1,2 +1,1 @@
-export { type NotificationQuestionsFormValues } from "./notification-questions-form-values.type";
 export { type RootStackParameterList } from "./root-stack-parameter-list.type";